// framework/ui/common/src/lib/model/menu-list/enum/class-list.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Enum defining all available menu item class list.
 */
export enum ClassList {

	/**
	 * Enum defines class type divider.
	 */
	divider = 'divider',
    /**
	* Enum defines class type collapsable.
	 */
	collapsable = 'collapsable',
    /**
	 * Enum defines class type dropdown btn
	 */
	dropdown = 'dropdown-btn-',
    /**
	 * Enum defines class type sublist.
	 */
	sublist = 'sublist',
    /**
	 * Enum defines class type fix
	 */
	fix = 'fix',
	/**
	 * Enum defines class type dropdown-toggle
	 */
	dropdownToggle ='dropdown-toggle',
	/**
	 * Enum defines class type dropdown-caret
	 */
	dropdownCaret = ' dropdown-caret',	
	/**
	 * Enum defines class type active
	 */
	active = 'active',
	/**
	 * Enum defines class type tlb-icons
	 */
	tlbIcons = 'tlb-icons',
	/**
	 * Enum defines class type menu-button
	 */	
	menuButton = 'menu-button',
	/**
	 * Enum defines class type ico-menu
	 */	
	icoMenu = 'ico-menu',

	/**
	 * Enum defines class type showimages
	 */
	showimages='showimages',

	/**
	 * Enum defines class type radio group
	 */
	radioGroup='radio-group',

	/**
	 * Enum defines class for popup
	 */
	popup = 'popup-menu',

	/**
	 * Enum defines class for overflow
	 */
	overflow = 'overflow',

	/**
	 * Enum defines class for container title
	 */
	title = 'title',

	/**
	 * Enum defines class for container title
	 */
	ellipsis = 'ellipsis'
}